venueflow-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
websockets==12.0
pydantic==2.6.4
google-generativeai==0.4.1
python-dotenv==1.0.1
aiofiles==23.2.1
